+++ webpack.config.js
@@ -1,10 +1,6 @@
 const path = require("path");
 const webpack = require("webpack");
-const HtmlPlugin = require("html-webpack-plugin");
-const MiniCssExtractPlugin = require("mini-css-extract-plugin");
-const CopyPlugin = require("copy-webpack-plugin");
 const envify = require("process-envify");
-require("babel-polyfill")
 const env = require("./env");
 // const pkg = require('./package');
 
@@ -15,7 +11,7 @@ module.exports = ({ prod = false } = {}) => ({
   target: 'node',
   mode: prod ? "production" : "development",
   context: SOURCE_ROOT,
-  entry: ['babel-polyfill', "./extension.js"],
+  entry: "./extension.js",
   output: {
     path: DISTRIBUTION_ROOT,
     // filename: "main.js",
@@ -33,23 +29,6 @@ module.exports = ({ prod = false } = {}) => ({
   module: {
     rules: [
       {
-        test: /\.svelte$/,
-        // exclude: /node_modules/,
-        use: {
-          loader: "svelte-loader",
-          options: {
-            emitCss: true,
-            hotReload: true,
-            hotOptions: {
-              // will display compile error in the client, avoiding page
-              // reload on error
-              optimistic: true
-            }
-          }
-        }
-      },
-
-      {
         test: /\.js$/,
         exclude: /node_modules/,
         use: [{
@@ -59,40 +38,7 @@ module.exports = ({ prod = false } = {}) => ({
           }
         }]
       },
-      {
-        test: /\.(sa|sc|c)ss$/,
-        // exclude: /node_modules/,
-        use: [
-          prod ? MiniCssExtractPlugin.loader : { loader: "style-loader" },
-          { loader: "css-loader", options: { importLoaders: 1, sourceMap: true } },
 
-
-        ]
-      },
-      {
-        test: /\.(png|jpe?g|gif|svg)(\?.*)?$/,
-        loader: "url-loader",
-        options: {
-          limit: 10000,
-          name: path.posix.join("assets", "images/[name].[hash].[ext]")
-        }
-      },
-      {
-        test: /\.(mp4|webm|ogg|mp3|wav|flac|aac)(\?.*)?$/,
-        loader: "url-loader",
-        options: {
-          limit: 10000,
-          name: path.posix.join("assets", "medias/[name].[hash].[ext]")
-        }
-      },
-      {
-        test: /\.(woff2?|eot|ttf|otf)(\?.*)?$/,
-        loader: "url-loader",
-        options: {
-          limit: 10000,
-          name: path.posix.join("assets", "fonts/[name].[hash].[ext]")
-        }
-      }
     ].filter(Boolean)
   },
   resolve: {
@@ -102,90 +48,29 @@ module.exports = ({ prod = false } = {}) => ({
     }
   },
   plugins: [
-    new MiniCssExtractPlugin({
-      filename: "style.[contenthash].css"
-    }),
+
     new webpack.DefinePlugin(envify(env)),
-    // new HtmlPlugin({
-    //   template: "index.html",
-    //   // minify: prod && {
-    //   //   removeComments: true,
-    //   //   collapseWhitespace: true,
-    //   //   removeAttributeQuotes: true
-    //   // },
-    //   // chunksSortMode: prod ? "manual" : "auto"
-    // }),
-    // new ScriptExtHtmlPlugin({
-    //   defaultAttribute: "defer",
-    //   preload: {
-    //     test: /\.js$/,
-    //     chunks: "initial"
-    //   },
-    //   prefetch: {
-    //     test: /\.js$/,
-    //     chunks: "all"
-    //   }
-    // }),
-    new CopyPlugin([
-      {
-        from: "./repl/**/*",
-        to: DISTRIBUTION_ROOT,
-        ignore: ["assets/styles/**/*"]
-      }
-    ]),
-    // !prod && new webpack.HotModuleReplacementPlugin(),
-    // prod && new webpack.optimize.AggressiveSplittingPlugin()
-    // prod && new GenerateSW({
-    //   exclude: [/\.(?:png|jpg|jpeg|svg)$/],
-    //   skipWaiting: true,
-    //   clientsClaim: true,
-    //   runtimeCaching: [
-    //     {
-    //       urlPattern: /\.(?:png|jpg|jpeg|svg)$/,
-    //       handler: 'cacheFirst',
-    //     },
-    //     {
-    //       urlPattern: new RegExp(env.SITE_URL),
-    //       handler: 'staleWhileRevalidate',
-    //       options: {
-    //         cacheableResponse: {
-    //           statuses: [0, 200],
-    //         },
-    //       },
-    //     },
-    //   ],
-    //   navigateFallback: '/',
-    //   navigateFallbackWhitelist: [/^(?!\/__).*/],
-    //   cacheId: pkg.name,
-    // }),
-    // prod && new PurgecssPlugin({
-    //   paths: glob.sync([
-    //     path.join(SOURCE_ROOT, './app/**/*.vue'),
-    //   ]),
-    //   whitelist: ['html', 'body'],
-    // }),
-    // prod && new RobotstxtPlugin(),
-    // prod && new SitemapPlugin(env.SITE_URL, [{ path: '/' }]),
+
   ].filter(Boolean),
-  // optimization: {
-  //   splitChunks: {
-  //     cacheGroups: {
-  //       common: {
-  //         name: "common",
-  //         chunks: "initial",
-  //         minChunks: 2
-  //       },
-  //       vendor: {
-  //         test: /[\\/]node_modules[\\/]/,
-  //         name: "vendor",
-  //         chunks: "all"
-  //       }
-  //     }
-  //   },
-  //   runtimeChunk: {
-  //     name: "manifest"
-  //   }
-  // },
+  optimization: {
+    splitChunks: {
+      cacheGroups: {
+        common: {
+          name: "common",
+          chunks: "initial",
+          minChunks: 2
+        },
+        vendor: {
+          test: /[\\/]node_modules[\\/]/,
+          name: "vendor",
+          chunks: "all"
+        }
+      }
+    },
+    runtimeChunk: {
+      name: "manifest"
+    }
+  },
   devServer: {
     contentBase: DISTRIBUTION_ROOT,
     historyApiFallback: true,